drivers/input: ft5x06: Handle more framebuffer hints
The driver currently only tries to honor doze mode by watching
for one of the low power mode hints and assuming all other modes
are not doze.  Recent changes to Ambient display have taken
advantage of additional power states, which we wrongly assumed
should not be using the touchpanel.  Fix this by handling all of
the currently supported power states.

* Original author Pat Erley <perley@cyngn.com>

Change-Id: If976ddbc30761fa3bb973f25ba66a8a7e3fa29dd